services/worker_service/requirements.txt
django==4.2
djangorestframework==3.15
psycopg[binary]
dj-database-url==2.1.0
python-dotenv==1.0.0
pika==1.3.2
requests==2.31.0
gunicorn==21.2.0
django-cors-headers==4.3.1
PyJWT==2.8.0
